added smoothing controller
