--- webapp/templates/stat/exp_smoothing.html
@@ -1,20 +1,0 @@
-{% extends "bootstrap.html" %}
-{% block content %}
-
-<div id="myPlot" style="width:100%;max-width:2000px;max-height:800px"></div>
-
-<script>
-// Define Data
-var data = [
-  {x: {{ time }} , y: {{ y }} , mode: 'lines', name: 'Number of sunspots'},
-  {x: {{ time }} , y: {{ y2 }} , mode: 'lines', name: 'Exponential smoothing (alpha=0.25)'},
-  {x: {{ time }} , y: {{ y3 }} , mode: 'lines', name: 'Double exponential smoothing (alpha=0.2, beta=0.2)'}
-];
-
-var layout = {title: "Graph of the exponential smoothing of the sunspots numbers for the last 100 years", xaxis: {title: "year", showgrid: true}, yaxis: {title: 'number of sunspots'}};
-
-// Display using Plotly
-Plotly.newPlot("myPlot", data, layout);
-</script>
-
-{% endblock %}
