--- webapp/templates/stat/triple_smoothing.html
@@ -1,19 +1,0 @@
-{% extends "bootstrap.html" %}
-{% block content %}
-
-<div id="myPlot" style="width:100%;max-width:2000px;max-height:800px"></div>
-
-<script>
-// Define Data
-var data = [
-  {x: {{ time }} , y: {{ y }} , mode: 'lines', name: 'Number of sunspots'},
-  {x: {{ time }} , y: {{ y2 }} , mode: 'lines', name: 'Triple exponential smoothing'},
-];
-
-var layout = {title: "Graph of the exponential smoothing of the sunspots numbers for 173 years", xaxis: {title: "year", showgrid: true}, yaxis: {title: 'number of sunspots'}};
-
-// Display using Plotly
-Plotly.newPlot("myPlot", data, layout);
-</script>
-
-{% endblock %}
